Add support for build description files in subdirectories
This has two side effects:
1. Generated outputs are now in subdirectories under the location
   of the source, not under the source tree root.
This was something I'd planned to around when this feature was
implemented, but it fell out as a side effect of cleanly handling
the leading "./" the subdirectory handling used.
It will probably be re-done at some point in the future to reflect
varying opinions about where build outputs should go.

2. Modules that link against libraries elsewhere in the build tree
   need to use '..' to find them.
This doesn't break things, but it's ugly enough that it will need
to be fixed up before the subdirectory support is declared done.