Fix handling of sources/intermediates
A "source" is something that exists in the build tree before we start
and has a name in srcnames[].  An "intermediate" is something that we
generate as part of the build process and then use as an input to a
subsequent step.

All exports are intermediates.

We currently only use intermediates as link inputs.  (This is more
consistently broken than the previous handling; it's still broken in a way
we don't exercise.)