Move to C99
Despite MSVC's long refusal to properly support it, Win64 makes it
impossible for a conforming C90 compiler to exist (long is fixed at 32
bits and size_t must be at least 64, breaking C90's guarantee that long
was the widest of the standard integer types).
So if we want to support Win64, we need to leave C90 (and its universal
support) behind.